Add a corner button to the globe
The button load async the icon used.
Important note: createjs mouse event management was confused
when using nested containers. I removed the main container and add the
globes to the stage.

There are a note here [1] saying: "In EaselJS 0.7.0, the mouseEnabled property will not work properly with nested Containers. Please check out the latest NEXT version in GitHub for an updated version with this issue resolved. The fix will be provided in the next release of EaselJS."

On version 0.8.1 still there are problems, not with the use of mouseEnabled,
but the events are sent to the wrong containers.

[1] http://createjs.com/docs/easeljs/classes/DisplayObject.html