Add information in the json to the objects, and method to validate
Added all the remaining data in the json to the ComicBox and Globes
objects properties, a method to transfor the properties to json
to be used in the persistence, and a method to test that the data is ok.
A constant DEBUG is defined and when true, the testing is done and logued.