Show the selected background image
The image data is loaded as data:image/png
and stored in the json file. Then when we read the image name,
we need check if is a file name or if is data.
This change is not supported by the python version yet,
but we can modify to make it compatible with newer versions.
The images are not scaled.