Consolidae images management
Cleanup. Before, thhe 'image_name' field could have a image_name
or a data uri depending on if the data was loaded from a .fototoon file
or if the image was added.
Now all the data uris go to the 'images' dictionary, andd 'image_name'
has always a name. When a image is added, a automatic name with a counter
is created.