Init sorting function
Add a button to switch to sorting mode, if there are moretwo boxes.
When switch mode, the canvas is resized.
Added a array 'previews' to the model data, to be used in the sorting
operation. As the previews re created when the user see a box,
we will need use the background images, without boxes, in the case
didn't displayed a box.
This patch only switch modes, do not implement sorting.