Solve error creating previews
The canvas is tainted when we display images loaded form disk
as the corner buttons on the globes, then we can't get the image
data using canvas.toDataUrl()

The solution is create a temporary canvas and use another ComicBox to
draw in that canvas. As the globes are not selected by default,
the controls are not displayed, then we avoid the tainted problem.