Initialize previews to show the globes for every box
Previusly we displayed the background image if a preview
was not available, and created the previews after a user leave a page
or before the user changed to the sort mode.
The problem with that was that if a user didn't changed to a given page
only the background would be displayed, without the globes.
Now, we check before change to sort mode, and create the pending
preview images if needed.