Save images on Android
The procedure used to save files is not working on Android.
To solve the problem in the case of the images, we added a plugin
to cordova (Canvas2Image) and use it in this patch.
We need add a Alert to notify the user the image is available
in the image gallery.